requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
pytz==2024.1
python-dateutil==2.9.0.post0
